move row stuff into row/
